adding separate screens and global store value context
