+++ src/index.js
@@ -1,11 +1,14 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+import React from "react";
+import ReactDOM from "react-dom";
+import "./index.css";
+import App from "./App";
+import GlobalStore from "./store/GlobalStore";
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <GlobalStore>
+      <App />
+    </GlobalStore>
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
