+++ src/store/GlobalStore.jsx
@@ -1,0 +1,8 @@
+import * as React from "react";
+import WeatherStore from "./WeatherContext";
+
+const GlobalStore = ({ children }) => {
+  return <WeatherStore>{children}</WeatherStore>;
+};
+
+export default GlobalStore;
